gpio: realtek-otto: Support per-cpu interrupts
On SoCs with multiple cores, it is possible that the GPIO interrupt
controller supports assigning specific pins to one or more cores.

This patch adds retrieval of the relevant register space from the
device's devicetree, but always assigns all interrupts to the first
core. Full support for CPU affinity will be implemented later.

Signed-off-by: Sander Vanheule <sander@svanheule.net>