irqchip/realtek-rtl: add CPU affinity support
CPU affinity is configured by setting the routing register values to the
required output, for all CPUs where the IRQ should be enabled. The GIM
register (interrupt enable) is always set or cleared for all CPUs to not
have the CPU affinity in two places.

This enables RTL839x and RTL930x SoC to balance IRQs over both available
VPEs.

TODO With this patch, not all parent interrupts can be used on VPE1.
This is because in arch/mips/kernel/smt-mt.c, only IP6 and IP7 are
enabled on VPE1. Adding a chained handler to the other interrupts does
not enable them at a later point for VPE1, so this needs to be fixed
somehow.