leds: add RTL8380 switch port LED driver
RTL8380 switch SoCs can control up to 3×28 LEDs to display status of the
switch's ethernet port. This driver allows to address these LEDs and
provides direct control, blink offloading, and switch port status
offloading.

Since offloading of the generic netdev trigger does not exist, this
driver provides a private trigger which achieves the same and is named
"realtek-switchport". Each LED will have a file name "rtl_hw_trigger" in
/sys/class/leds/$LED, where the requested trigger mode can be written
to. If an unsupported mode is requested, this will be reported to the
user. When the trigger is then activated, the LED will be added to a
group of LEDs with the same trigger conditions. If it is not possible to
add the LED to a group, the user must use an already existing trigger
condition, or use direct LED control with a software trigger.

Some common modes (i.e. values for "rtl_hw_trigger") are:
  - Link present, blink on activity: 1f
  - 100M/10M link, blink on activity: f
  - 1G link present: 10

Signed-off-by: Sander Vanheule <sander@svanheule.net>