hwmon: gpio-fan: convert control gpios to array
Instead of open-coding a gpio array, convert the driver to the gpiod
array interface for the fan speed control GPIOs. This reduces the
complexity of the code a bit.

Since the gpio-fan DT binding specifies that the control gpios are
listed MSB-first, the gpio_desc array needs to be reversed to allow
use of the gpiod array API.

Signed-off-by: Sander Vanheule <sander@svanheule.net>